Norminette

+++ main.c
@@ -3,9 +3,9 @@
 
 int main()
 {
-	char *o = "or: %llu\n";
-	char *f = "ft: %llu\n";
-	unsigned long long n = (unsigned long long)-12345612220;
+	char *o = "or: %-+05.5lli\n";
+	char *f = "ft: %-+05.5lli\n";
+	long long n = -12345612220;
 	int a = printf(o, n);
 	int b = ft_printf(f, n);
 	printf("%i %i\n", a, b);
